requirements.txt
bcrypt==3.2.2
Flask==2.1.3
Flask-RESTful==0.3.9
SQLAlchemy==1.4.39
pandas==1.4.3
psycopg2==2.9.3
redis==4.3.4
requests==2.28.1
passlib==1.7.4
python-dotenv==0.17.1
python-jose==3.3.0